Fixed the sharpen filter. The methodology I had approached the problem with was incorrect. The process is not to subtract a scaled invert of the coarse from the normal.
The correct process is to acquire a "Fine" overlay by subtracting the
"Coarse" output from the standard input, then to scale that "Fine"
overlay down and to add it (with a scaling factor) to the original.

I will have to consider my approach for adjusting the intensity of the
sharpness. Varying the strength of the guassian blur is an option - but
I'm not sure that will really change the intensity. Altering the scaling
of the "Fine" overlay is another option, but it can easily cause the
image to brighten.

I think I will work on a contrast filter next to better understand how
brightness and contrast interact in image editors.